# api/openapi/reimbursement-v1.yaml
openapi: 3.1.0
info:
  title: AI Reimbursement API
  version: 1.0.0
paths:
  /api/v1/admin/policies/publish:
    post:
      operationId: publishPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, effectiveDate]
              properties:
                id: { type: string }
                effectiveDate: { type: string, format: date }
      responses:
        "201": { description: Policy published }
        default: { $ref: "#/components/responses/Error" }
...
components:
...
  responses:
    Error:
      description: Safe business error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string